backend/email_collector/requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.110.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
supabase>=2.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
typing-extensions>=4.8.0